+++ business-services/billing-service/src/main/java/org/egov/demand/util/Constants.java
@@ -20,6 +20,8 @@ public class Constants {
 	public static final String TAXPERIOD_PATH_CODE = "$.MdmsRes.BillingService.TaxPeriod";
 	public static final String TAXHEADMASTER_PATH_CODE = "$.MdmsRes.BillingService.TaxHeadMaster";
 	public static final String BUSINESSSERVICE_PATH_CODE = "$.MdmsRes.BillingService.BusinessService.*.code";
+	public static final String BUSINESSSERVICE_MODULE_PATH = "$.MdmsRes.BillingService.BusinessService";
+	
 	
 	public static final String TAXPERIOD_MASTERNAME = "TaxPeriod";
 	public static final String TAXHEAD_MASTERNAME = "TaxHeadMaster";
@@ -137,8 +139,20 @@ public class Constants {
 	public static final String PAYMENT_ID_KEY = "paymentId";
 
 	public static final String EMPLOYEE_TYPE_CODE = "EMPLOYEE";
+	
+	
+	/*
+	 * Amendment Constants
+	 */
+	
+	public static final String NOTETYPE_REPLACE_STRING = "NOTETYPE";
 
-
+	public static final String CONSUMERCODE_REPLACE_STRING = "CONSUMERCODE";
+	
+	public static final String CREDIT_NOTE_VALUE = "CN";
+	
+	public static final String DEBIT_NOTE_VALUE = "DN";
+	
 
 	private Constants() {}
 }