+++ business-services/egov-hrms/src/main/java/org/egov/hrms/web/contract/User.java
@@ -51,7 +51,7 @@ public class User {
 
     @NotNull
     @Size(max=100)
-    @Pattern(regexp = "^[a-zA-Z0-9 \\-'`\\.]*$", message = "Invalid name. Only alphabets and special characters -, ',`, .")
+    @Pattern(regexp = "^[a-zA-Z \\-'`\\.]*$", message = "Invalid name. Only alphabets and special characters -, ',`, .")
     @JsonProperty("name")
     private String name;