+++ business-services/collection-services/CHANGELOG.md
@@ -3,6 +3,9 @@
 # Changelog
 All notable changes to this module will be documented in this file.
 
+## 1.1.3 - 2021-02-26
+- Updated domain name in application.properties
+
 ## 1.1.2 - 2021-01-12
 - Query Optimized
 - Payment search/workflow path param based URL config added