+++ business-services/egov-apportion-service/CHANGELOG.md
@@ -3,6 +3,10 @@
 # Changelog
 All notable changes to this module will be documented in this file.
 
+## 1.1.4 2021-02-26
+
+- Updated domain name in application.properties
+
 ## 1.1.3 2021-01-12
 
 - Apportion and backupdate to demands enabled for zero payments, which was blocked till this point of time