+++ business-services/egov-hrms/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 All notable changes to this module will be documented in this file.
 
+## 1.2.1 - 2021-02-26
+- Updated domain name in application.properties
+
 ## 1.2.0 - 2021-01-12
 - Added employee reactivation feature